minor outbreak issue #1

+++ kill.sh
@@ -1,14 +1,11 @@
 #!/bin/sh
-
-# TODO: grep for the Bad Error keyword
-
-# to kill the PID
-for i in `ls *.pid`
-do
-    for j in `cat ${i}`
+a=`grep "^\[ERROR\]" log.txt`
+if [ $? -eq 0 ]; then
+    for j in `cat log.pid`
     do
-        kill -9 ${j}
+      kill -9 ${j}
     done
-    rm ${i}
-done
-
+    rm log.pid
+    rm log.txt
+    bash run.sh
+fi
